Description: codex clean. Feature or Bugfix:Feature Binary Source: No

+++ core/authentication/src/auth_connection.c
@@ -42,7 +42,7 @@ typedef struct {
     int32_t connModule;
 } PostDataInfo;
 
-static int32_t PostDataByConn(const PostDataInfo *info, char *buf, uint32_t postDataLen)
+static int32_t PostDataByConn(const PostDataInfo *info, char *buf, int32_t postDataLen)
 {
     int64_t seq = 0;
     if (info->connModule == MODULE_DEVICE_AUTH ||
@@ -62,7 +62,7 @@ static int32_t PostDataByConn(const PostDataInfo *info, char *buf, uint32_t post
     postParam.flag = info->side;
     postParam.pid = 0;
     postParam.buf = buf;
-    postParam.len = postDataLen + ConnGetHeadSize();
+    postParam.len = postDataLen + (int32_t)ConnGetHeadSize();
 
     return ConnPostBytes(info->connectionId, &postParam);
 }
@@ -144,7 +144,7 @@ int32_t AuthPostData(const AuthDataHead *head, const uint8_t *data, uint32_t len
         SoftBusLog(SOFTBUS_LOG_AUTH, SOFTBUS_LOG_INFO,
             "auth start post data, authId is %lld, connectionId is %u, moduleId is %d, seq is %lld",
             auth->authId, info.connectionId, info.connModule, info.seq);
-        if (PostDataByConn(&info, connPostData, postDataLen) != SOFTBUS_OK) {
+        if (PostDataByConn(&info, connPostData, (int32_t)postDataLen) != SOFTBUS_OK) {
             SoftBusLog(SOFTBUS_LOG_AUTH, SOFTBUS_LOG_ERROR, "PostDataByConn failed");
             return SOFTBUS_ERR;
         }
@@ -363,7 +363,7 @@ void AuthSendCloseAck(uint32_t connectionId)
     info.seq = 0;
     info.connectionId = connectionId;
     info.connModule = MODULE_DEVICE_AUTH;
-    if (PostDataByConn(&info, connPostData, postDataLen) != SOFTBUS_OK) {
+    if (PostDataByConn(&info, connPostData, (int32_t)postDataLen) != SOFTBUS_OK) {
         SoftBusLog(SOFTBUS_LOG_AUTH, SOFTBUS_LOG_ERROR, "PostDataByConn failed");
         return;
     }
